fix: remove redundant GLU, add cmath include and change many math functions, and add GLM dependency in README
Many function of float versions are replaced with the standard version.
Although this can produce conversion error, these are most likely
useless since they belongs to the CPU shader part, which is not in use.
This change is necessary because acoording to [1], in some version of
GCC, the float and long double counterpart of <cmath> functions are not
actually defined, which leads to compilation error.

[1] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=79700